% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cv_poet}
\alias{cv_poet}
\title{Cross-Validation of POET (non-parallel, Cpp version)}
\usage{
cv_poet(Y, K_seq, C_seq, thres = "soft", matrix = "cor", kfolds = 5L)
}
\arguments{
\item{Y}{p by n data matrix.}

\item{K_seq}{sequence of K for cv. K is the number of latent factor.}

\item{C_seq}{sequence of C for cv. C is the shrinkage intensity.}

\item{thres}{shrinkage method.}

\item{matrix}{indicator of shrinkage matrix, "cor" (default) or "vad".}

\item{kfolds}{kfolds-cv.}
}
\value{
A list with two elements
\item{params}{params expanded by K_seq and C_seq.}
\item{cri}{the calculated criteria across the params using kfolds-cv.}
\item{K_opt}{the optimal K over the K_seq.}
\item{C_opt}{the optimal C over the C_seq.}
}
\description{
Cross-Validation criterion is the Kullback–Leibler loss (KLL) function. Note
that the n observations are partitioned in order (not randomly), i.e. if n =
100, 5-folds cv will use the partitions as follows: [1, ..., 20], [21, ...,
40], [41, ..., 60], [61, ..., 80], [81, ..., 100].
}
\details{
Cpp version of \code{\link[poetREV]{cv_poet_par}}, see
\code{\link[poetREV]{cv_poet_par}} for examples.
}
\seealso{
\code{\link[poetREV]{cv_poet_par}}
}
