% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_poet.r
\name{cv_poet_par}
\alias{cv_poet_par}
\title{Cross-Validation of POET (parallel version)}
\usage{
cv_poet_par(Y, K_seq, C_seq, thres = "soft", matrix = "cor", kfolds = 5)
}
\arguments{
\item{Y}{p by n data matrix.}

\item{K_seq}{sequence of K for cv. K is the number of latent factor.}

\item{C_seq}{sequence of C for cv. C is the shrinkage intensity.}

\item{thres}{shrinkage method.}

\item{matrix}{indicator of shrinkage matrix, "cor" (default) or "vad".}

\item{kfolds}{kfolds-cv.}
}
\value{
A list with two elements
\item{params}{params expanded by K_seq and C_seq.}
\item{cri}{the calculated criteria across the params using kfolds-cv.}
\item{K_opt}{the optimal K over the K_seq.}
\item{C_opt}{the optimal C over the C_seq.}
}
\description{
Cross-Validation criterion is the Kullback–Leibler loss (KLL) function. Note
that the n observations are partitioned in order (not randomly), i.e. if n =
100, 5-folds cv will use the partitions as follows: [1, ..., 20], [21, ...,
40], [41, ..., 60], [61, ..., 80], [81, ..., 100].
}
\details{
If you want to split the data in a random way, you can simply modify this
function to meet your requirement.
}
\examples{
n <- 50; p <- 100; k <- 3
set.seed(1)
B <- matrix(rnorm(100 * 3), 100, 3)
F <- MASS::mvrnorm(n, rep(0, k), diag(1, k))
U <- MASS::mvrnorm(n, rep(0, p), diag(1, p))
Y <- t(F \%*\% t(B) + U)
res_cv1 <- poetREV::cv_poet(Y, 1:5, seq(.5, 1.5, .1), "soft", "cor", kfolds = 10)

library(doParallel, quietly = TRUE)
## cl <- makeCluster(detectCores()-1)
cl <- makeCluster(2) ## So funny... https://stackoverflow.com/questions/15648772
registerDoParallel(cl)
res_cv2 <- poetREV::cv_poet_par(Y, 1:5, seq(.5, 1.5, .1), "soft", "cor", kfolds = 10)
stopCluster(cl)

res_cv1$params - res_cv2$params
res_cv1$cri - res_cv2$cri

## We can obtain K by Information Criteria first, then cv the C_seq
## To reduce the computational burden
K_opt <- floor(mean(poetREV::poet_Khat(Y)))
res_cv3 <- poetREV::cv_poet(Y, K_opt, seq(.5, 1.5, .1), "soft", "cor", kfolds = 10)

}
\references{
Fan, J., Liao, Y., & Mincheva, M. (2013). Large covariance estimation by thresholding principal orthogonal complements. Journal of the Royal Statistical Society Series B: Statistical Methodology, 75(4), 603-680.
}
\seealso{
\code{\link[poetREV]{cv_poet}}
}
