% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_poet.r
\name{poet2}
\alias{poet2}
\title{The R version of \code{\link[poetREV]{poet}}, which is written in Cpp.}
\usage{
poet2(Y, K = -Inf, C = -Inf, thres = "soft", matrix = "cor")
}
\arguments{
\item{Y}{p by n data matrix.}

\item{K}{the number of latent factors.}

\item{C}{the shrinkage intensity.}

\item{thres}{shrinkage method. Default is "soft".}

\item{matrix}{indicator of shrinkage matrix, "cor" (default) or "vad".}
}
\value{
a list
}
\description{
This function is written for the purpose of checking the correctness of
\code{\link[poetREV]{poet}}, see the examples in
\code{\link[poetREV]{poet}}.
}
