% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_poet.r
\name{poet_Cmin}
\alias{poet_Cmin}
\title{Cmin - Minimum threshold constant}
\usage{
poet_Cmin(Y, K, thres, matrix)
}
\arguments{
\item{Y}{p by n matrix of raw data, where p is the dimensionality, n is the
sample size. It is recommended that Y is de-meaned, i.e., each row has
zero mean.}

\item{K}{number of factors. K is pre-determined by the users. Suggestions on
    choosing K:

(1) A simple way of determining K is to count the number of very spiked
(much larger than others) eigenvalues of the p by p sample covariance matrix
of Y.

(2) A formal data-driven way of determining K is described in Bai and Ng
(2002):"Determining the number of factors in approximate factor models",
Econometrica, 70, 191-221. This procedure requires a one-dimensional
optimization.

(3) POET is very robust to over-estimating K. But under-estimating K can
result to VERY BAD performance. Therefore we strongly recommend choosing a
relatively large K (normally less than 8) to avoid missing any important
common factor.

(4) K=0 corresponds to threshoding the sample covariance directly.}

\item{thres}{choice of thresholding. Users can choose from three
    thresholding methods:

'soft': soft thresholding

'hard': hard thresholding

'scad': scad thresholding

'alasso': adaptive lasso thresholding

Details are found in Rothman et al. (2009):
"Generalized thresholding of large covariance matrices." JASA, 104, 177-186}

\item{matrix}{the option of thresholding either correlation or covairance
    matrix. Users can choose from:

'cor': threshold the error correlation matrix then transform back to
covariance matrix

'vad': threshold the error covariance matrix directly.}
}
\value{
a scalar, the minimum threshold constant
}
\description{
(Copy from POET package) This function is for determining the minimum
constant in the threshold that guarantees the positive definiteness of the
POET estimator.
}
\details{
Model: Y_t=Bf_t+u_t, where B, f_t and u_t represent factor loading matrix,
common factors and idiosyncratic error respectively. Only Y_t is
observable. t=1,...,n. Dimension of Y_t is p. The goal is to estimate the
covariance matrices of Y_t and u_t.

Note: (1) POET is optimization-free, so no initial value, tolerant, or
maximum iterations need to be specified as inputs.

(2) We can apply the adaptive thresholding (Cai and Liu 2011, JASA) on
either the correlation matrix or the covariance matrix, specified by the
option 'matrix'.

(3) If no factor structure is assumed, i.e., no common factors exist and
var(Y_t) itself is sparse, set K=0.
}
\examples{
n <- 50; p <- 100; k <- 3
set.seed(1)
B <- matrix(rnorm(100 * 3), 100, 3)
F <- MASS::mvrnorm(n, rep(0, k), diag(1, k))
U <- MASS::mvrnorm(n, rep(0, p), diag(1, p))
Y <- t(F \%*\% t(B) + U)
poetREV::poet_Cmin(Y, 3, "soft", "cor")
## [1] 0.2605228
}
