% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poet_Khat}
\alias{poet_Khat}
\title{Determine the number of latent factors in POET (approximate factor model)}
\usage{
poet_Khat(Y, kmax = 15L)
}
\arguments{
\item{Y}{p by n matrix of raw data, where p is the dimensionality, n is the
sample size. It is recommended that Y is de-meaned, i.e., each row has
zero mean.}

\item{kmax}{the given upper bound of number of factors. The optimal k is
searched from 1 to kmax.}
}
\value{
A vector with two elements
\item{K1BN}{(The 1st element) estimated number of factors based on the first infomation criterion using Bai & Ng method}
\item{K2BN}{(The 2nd element) estimated number of factors based on the second infomation criterion using Bai & Ng method}
}
\description{
This function is for calculating the optimal number of factors in an
approximate factor model.
}
\details{
(Copy from POET package) This method was proposed by Bai & Ng (2002) and
Hallin & Liska (2007). They propose two penalty functions and in turn
minimize the corresponding information criteria. Notice that this method may
underestimate K. POET is very robust to over-estimating K. But
under-estimating K can result to VERY BAD performance. Therefore we strongly
recommend choosing a relatively large K (normally less than 8) to avoid
missing any important common factor.

IMPORTANT: the calculation of Hallin & Liska method (2007) is wrong in the
original POET package, here we only calculate the criteria from Bai & Ng
(2002).
}
\examples{
n <- 50; p <- 100; k <- 3
set.seed(1)
B <- matrix(rnorm(100 * 3), 100, 3)
F <- MASS::mvrnorm(n, rep(0, k), diag(1, k))
U <- MASS::mvrnorm(n, rep(0, p), diag(1, p))
Y <- t(F \%*\% t(B) + U)
poetREV::poet_Khat(Y)
poetREV::poet_Khat2(Y)
}
\references{
Hallin, M., & Liška, R. (2007). Determining the number of factors in the general dynamic factor model. Journal of the American Statistical Association, 102(478), 603-617.

Bai, J., & Ng, S. (2002). Determining the number of factors in approximate factor models. Econometrica, 70(1), 191-221.
}
