% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_poet.r
\name{poet_Khat2}
\alias{poet_Khat2}
\title{The R version of \code{\link[poetREV]{poet_Khat}}, which is written in Cpp.}
\usage{
poet_Khat2(Y)
}
\arguments{
\item{Y}{p by n data matrix.}
}
\value{
a vector with two elements.
}
\description{
This function is written for the purpose of checking the correctness of
\code{\link[poetREV]{poet_Khat}}, see the examples in
\code{\link[poetREV]{poet_Khat}}.
}
